# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Small Town Watchdog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /source-registry:
    get:
      operationId: listSourceRegistry
      tags: [source-registry]
      summary: List source registry entries
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceRegistry"
    post:
      operationId: createSourceRegistry
      tags: [source-registry]
      summary: Create a source registry entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceRegistryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRegistry"
  /source-registry/{id}:
    get:
      operationId: getSourceRegistry
      tags: [source-registry]
      summary: Get a source registry entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRegistry"
        "404":
          description: Not found
    patch:
      operationId: updateSourceRegistry
      tags: [source-registry]
      summary: Update a source registry entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceRegistryPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceRegistry"
        "404":
          description: Not found
    delete:
      operationId: deleteSourceRegistry
      tags: [source-registry]
      summary: Delete a source registry entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /source-submissions:
    get:
      operationId: listSourceSubmissions
      tags: [source-registry]
      summary: List community source submissions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status (pending, approved, rejected)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceSubmission"
    post:
      operationId: createSourceSubmission
      tags: [source-registry]
      summary: Submit a source for community review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceSubmissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceSubmission"
  /source-submissions/{id}:
    patch:
      operationId: updateSourceSubmission
      tags: [source-registry]
      summary: Approve or reject a source submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceSubmissionPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceSubmission"
        "404":
          description: Not found
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List state/county/parish locations
      parameters:
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state code (e.g. LA, MS)
        - name: county
          in: query
          schema:
            type: string
          description: Filter by county/parish name
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a single location by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          description: Not found
...
  /error-reports:
    post:
      operationId: createErrorReport
      tags: [error-reports]
      summary: Submit an error report for a civic item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErrorReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorReport"
  /admin/error-reports:
    get:
      operationId: listAdminErrorReports
      tags: [error-reports]
      summary: Admin — list all error reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorReport"
  /admin/error-reports/{id}:
    patch:
      operationId: updateErrorReport
      tags: [error-reports]
      summary: Admin — update error report status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErrorReportPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorReport"
        "404":
          description: Not found
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders for a device
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
          description: The user's device identifier
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
        "400":
          description: Missing deviceId

    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "400":
          description: Missing required fields
  /reminders/{id}:
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
...
components:
  schemas:
...
    SourceRegistry:
      type: object
      required: [id, state, entityName, entityType, sourceUrl, sourceCategory, sourcePlatform, verificationStatus, isActive, createdAt]
      properties:
        id:
          type: integer
        state:
          type: string
        county:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        entityName:
          type: string
        entityType:
          type: string
        sourceUrl:
          type: string
        sourceCategory:
          type: string
        sourcePlatform:
          type: string
        verificationStatus:
          type: string
        lastChecked:
          type: ["string", "null"]
        lastSuccessfulUpdate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    SourceRegistryInput:
      type: object
      required: [state, entityName, entityType, sourceUrl, sourceCategory, sourcePlatform]
      properties:
        state:
          type: string
        county:
          type: string
        city:
          type: string
        entityName:
          type: string
        entityType:
          type: string
        sourceUrl:
          type: string
        sourceCategory:
          type: string
        sourcePlatform:
          type: string
        verificationStatus:
          type: string
        notes:
          type: string
        isActive:
          type: boolean
    SourceRegistryPatch:
      type: object
      properties:
        state:
          type: string
        county:
          type: string
        city:
          type: string
        entityName:
          type: string
        entityType:
          type: string
        sourceUrl:
          type: string
        sourceCategory:
          type: string
        sourcePlatform:
          type: string
        verificationStatus:
          type: string
        notes:
          type: string
        isActive:
          type: boolean
        lastChecked:
          type: string
        lastSuccessfulUpdate:
          type: string
    SourceSubmission:
      type: object
      required: [id, state, city, entityName, entityType, sourceUrl, sourceCategory, status, createdAt]
      properties:
        id:
          type: integer
        state:
          type: string
        county:
          type: ["string", "null"]
        city:
          type: string
        entityName:
          type: string
        entityType:
          type: string
        sourceUrl:
          type: string
        sourceCategory:
          type: string
        submitterNote:
          type: ["string", "null"]
        status:
          type: string
        reviewNote:
          type: ["string", "null"]
        createdAt:
          type: string
    SourceSubmissionInput:
      type: object
      required: [state, city, entityName, entityType, sourceUrl, sourceCategory]
      properties:
        state:
          type: string
        county:
          type: string
        city:
          type: string
        entityName:
          type: string
        entityType:
          type: string
        sourceUrl:
          type: string
        sourceCategory:
          type: string
        submitterNote:
          type: string
    SourceSubmissionPatch:
      type: object
      properties:
        status:
          type: string
        reviewNote:
          type: string
...
    Location:
      type: object
      required: [id, stateCode, stateName, countyParish, countyType, isActive, coverageStatus, createdAt]
      properties:
        id:
          type: integer
        stateCode:
          type: string
        stateName:
          type: string
        countyParish:
          type: string
        countyType:
          type: string
        isActive:
          type: boolean
        coverageStatus:
          type: string
        createdAt:
          type: string
...
    ErrorReport:
      type: object
      required: [id, reportType, status, createdAt]
      properties:
        id:
          type: integer
        civicItemId:
          type: ["integer", "null"]
        reportType:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ErrorReportInput:
      type: object
      required: [reportType]
      properties:
        civicItemId:
          type: integer
        reportType:
          type: string
        message:
          type: string
    ErrorReportPatch:
      type: object
      properties:
        status:
          type: string
...
    Reminder:
      type: object
      required: [id, userDeviceId, reminderType, remindEarlyVoting, remindDayBefore, remindTwoHoursBefore, remindPollClosing, createdAt]
      properties:
        id:
          type: integer
        userDeviceId:
          type: string
        civicItemId:
          type: ["integer", "null"]
        reminderType:
          type: string
        remindEarlyVoting:
          type: boolean
        remindDayBefore:
          type: boolean
        remindTwoHoursBefore:
          type: boolean
        remindPollClosing:
          type: boolean
        createdAt:
          type: string
    ReminderInput:
      type: object
      required: [userDeviceId]
      properties:
        userDeviceId:
          type: string
        civicItemId:
          type: integer
        reminderType:
          type: string
        remindEarlyVoting:
          type: boolean
        remindDayBefore:
          type: boolean
        remindTwoHoursBefore:
          type: boolean
        remindPollClosing:
          type: boolean